[Misc] Update folder structures for the Note component
