+++ src/components/Note/ReplyArea/index.js
@@ -1,0 +1,3 @@
+import ReplyArea from './ReplyArea';
+
+export default ReplyArea;
