+++ src/components/NoteContent/ContentArea/index.js
@@ -1,0 +1,3 @@
+import ContentArea from './ContentArea';
+
+export default ContentArea;
