+++ src/components/NoteContent/NoteContent.js
@@ -1,19 +1,17 @@
 import React, {
   useState,
-  useRef,
   useEffect,
   useContext,
   useMemo,
   useCallback,
 } from 'react';
 import PropTypes from 'prop-types';
-import classNames from 'classnames';
 import { useDispatch, useSelector, shallowEqual } from 'react-redux';
 import { useTranslation } from 'react-i18next';
 import Autolinker from 'autolinker';
 import dayjs from 'dayjs';
 
-import AutoResizeTextarea from 'components/AutoResizeTextarea';
+import ContentArea from 'components/NoteContent/ContentArea';
 import NotePopup from 'components/NotePopup';
 import NoteContext from 'components/Note/Context';
 import Icon from 'components/Icon';
@@ -246,79 +244,3 @@ const NoteContent = ({ annotation }) => {
 NoteContent.propTypes = propTypes;
 
 export default NoteContent;
-
-// a component that contains the content textarea, the save button and the cancel button
-const ContentArea = ({
-  annotation,
-  setIsEditing,
-  textAreaValue,
-  onTextAreaValueChange,
-}) => {
-  const contents = annotation.getContents();
-  const [t] = useTranslation();
-  const textareaRef = useRef();
-
-  useEffect(() => {
-    // on initial mount, focus the last character of the textarea
-    if (textareaRef.current) {
-      textareaRef.current.focus();
-
-      const textLength = textareaRef.current.value.length;
-      textareaRef.current.setSelectionRange(textLength, textLength);
-    }
-  }, []);
-
-  const setContents = e => {
-    // prevent the textarea from blurring out which will unmount these two buttons
-    e.preventDefault();
-
-    const hasEdited = textAreaValue !== contents;
-    if (hasEdited) {
-      core.setNoteContents(annotation, textAreaValue);
-      if (annotation instanceof window.Annotations.FreeTextAnnotation) {
-        core.drawAnnotationsFromList([annotation]);
-      }
-
-      setIsEditing(false);
-    }
-  };
-
-  const saveBtnClass = classNames({
-    disabled: textAreaValue === contents,
-  });
-
-  return (
-    <div className="edit-content">
-      <AutoResizeTextarea
-        ref={el => {
-          textareaRef.current = el;
-        }}
-        value={textAreaValue}
-        onChange={onTextAreaValueChange}
-        onBlur={() => setIsEditing(false)}
-        onSubmit={setContents}
-        placeholder={`${t('action.comment')}...`}
-      />
-      <span className="buttons">
-        <button className={saveBtnClass} onMouseDown={setContents}>
-          {t('action.save')}
-        </button>
-        <button
-          onMouseDown={() => {
-            setIsEditing(false);
-            onTextAreaValueChange(contents);
-          }}
-        >
-          {t('action.cancel')}
-        </button>
-      </span>
-    </div>
-  );
-};
-
-ContentArea.propTypes = {
-  annotation: PropTypes.object.isRequired,
-  setIsEditing: PropTypes.func.isRequired,
-  textAreaValue: PropTypes.string,
-  onTextAreaValueChange: PropTypes.func.isRequired,
-};
