+++ src/components/NotesPanel/NormalList/index.js
@@ -1,0 +1,3 @@
+import NormalList from './NormalList';
+
+export default NormalList;
