+++ src/components/NotesPanel/NotesPanel.js
@@ -1,10 +1,10 @@
-import React, { useState, useRef, useEffect, useImperativeHandle } from 'react';
+import React, { useState, useRef, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { useSelector, shallowEqual } from 'react-redux';
-import Measure from 'react-measure';
-import { CellMeasurer, CellMeasurerCache, List } from 'react-virtualized';
 import { useTranslation } from 'react-i18next';
 
+import VirtualizedList from 'components/NotesPanel/VirtualizedList';
+import NormalList from 'components/NotesPanel/NormalList';
 import Dropdown from 'components/Dropdown';
 import Note from 'components/Note';
 import NoteContext from 'components/Note/Context';
@@ -284,133 +284,3 @@ const NotesPanel = ({ display }) => {
 NotesPanel.propTypes = propTypes;
 
 export default NotesPanel;
-
-const listPropTypes = {
-  notes: PropTypes.array.isRequired,
-  children: PropTypes.func.isRequired,
-  onScroll: PropTypes.func.isRequired,
-  initialScrollTop: PropTypes.number.isRequired,
-};
-
-const cache = new CellMeasurerCache({ defaultHeight: 50, fixedWidth: true });
-
-const VirtualizedList = React.forwardRef(
-  ({ notes, children, onScroll, initialScrollTop }, forwardedRef) => {
-    const listRef = useRef();
-    const [dimension, setDimension] = useState({ width: 0, height: 0 });
-
-    useImperativeHandle(forwardedRef, () => ({
-      scrollToPosition: scrollTop => {
-        listRef.current.scrollToPosition(scrollTop);
-      },
-      scrollToRow: index => {
-        listRef.current.scrollToRow(index);
-      },
-    }));
-
-    useEffect(() => {
-      listRef.current.scrollToPosition(initialScrollTop);
-    }, [initialScrollTop]);
-
-    const _resize = index => {
-      cache.clear(index);
-      listRef.current?.recomputeRowHeights(index);
-    };
-
-    const handleScroll = ({ scrollTop }) => {
-      onScroll(scrollTop);
-    };
-
-    /* eslint-disable react/prop-types */
-    const rowRenderer = ({ index, key, parent, style }) => {
-      const currNote = notes[index];
-
-      return (
-        <CellMeasurer
-          key={`${key}${currNote.Id}`}
-          cache={cache}
-          columnIndex={0}
-          parent={parent}
-          rowIndex={index}
-        >
-          <div style={style}>
-            {children(notes, index, () => _resize(index))}
-          </div>
-        </CellMeasurer>
-      );
-    };
-
-    return (
-      <Measure bounds onResize={({ bounds }) => setDimension(bounds)}>
-        {({ measureRef }) => (
-          <div ref={measureRef} className="virtualized-notes-container">
-            <List
-              deferredMeasurementCache={cache}
-              style={{ outline: 'none' }}
-              height={dimension.height}
-              width={dimension.width}
-              overscanRowCount={10}
-              ref={listRef}
-              rowCount={notes.length}
-              rowHeight={cache.rowHeight}
-              rowRenderer={rowRenderer}
-              onScroll={handleScroll}
-            />
-          </div>
-        )}
-      </Measure>
-    );
-  },
-);
-
-VirtualizedList.propTypes = listPropTypes;
-
-const NormalList = React.forwardRef(
-  ({ notes, children, onScroll, initialScrollTop }, forwardedRef) => {
-    const listRef = useRef();
-
-    useImperativeHandle(forwardedRef, () => ({
-      scrollToPosition: scrollTop => {
-        listRef.current.scrollTop = scrollTop;
-      },
-      scrollToRow: index => {
-        const parent = listRef.current;
-        const child = parent.children[index];
-
-        const parentRect = parent.getBoundingClientRect();
-        const childRect = child.getBoundingClientRect();
-
-        const isViewable =
-          childRect.top >= parentRect.top &&
-          childRect.top <= parentRect.top + parent.clientHeight;
-        if (!isViewable) {
-          parent.scrollTop = childRect.top + parent.scrollTop - parentRect.top;
-        }
-      },
-    }));
-
-    useEffect(() => {
-      listRef.current.scrollTop = initialScrollTop;
-    }, [initialScrollTop]);
-
-    const handleScroll = e => {
-      onScroll(e.target.scrollTop);
-    };
-
-    return (
-      <div
-        ref={listRef}
-        className="normal-notes-container"
-        onScroll={handleScroll}
-      >
-        {notes.map((currNote, index) => (
-          <React.Fragment key={`${index}${currNote.Id}`}>
-            {children(notes, index)}
-          </React.Fragment>
-        ))}
-      </div>
-    );
-  },
-);
-
-NormalList.propTypes = listPropTypes;
