+++ src/components/NotesPanel/VirtualizedList/index.js
@@ -1,0 +1,3 @@
+import VirtualizedList from './VirtualizedList';
+
+export default VirtualizedList;
